Added the post action for report, and fixed a few sql errors
git-svn-id: http://nolll.homeip.net:8081/svn/PokerBunch/trunk@121 fbe6c1e3-6fed-8744-b610-9659f20f9f49

+++ Infrastructure.Tests/Factories/CashgameFactoryTests.cs
@@ -49,7 +49,7 @@ namespace Infrastructure.Tests.Factories{
         [Test]
 		public void Get_CashgamePropertiesAreSet(){
 			var sut = GetSut();
-			var result = sut.Create(_location, _status, _id, _results);
+			var result = sut.Create(_location, (int)_status, _id, _results);
 
 			Assert.AreEqual(_location, result.Location);
             Assert.AreEqual(_status, result.Status);
